condensed print document architecture and design
