--- website/src/pages/documents/print-layout.css.ts
@@ -1,119 +1,0 @@
-import { globalStyle, style } from "@vanilla-extract/css"
-import { mediaQueries, vspace } from "src/style/constants"
-
-const printFonts = {
-  quattrocento: `"Quattrocento", serif`,
-}
-
-const printFontSizes = {
-  large: "1.1rem",
-  legendItem: "1rem",
-}
-
-export const printHeading = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      fontFamily: printFonts.quattrocento,
-      fontSize: printFontSizes.large,
-      fontWeight: 700,
-      lineHeight: 1,
-      marginBottom: 0,
-    },
-  },
-})
-
-export const printBreadcrumbs = style({
-  "@media": {
-    [mediaQueries.print]: {
-      fontFamily: printFonts.quattrocento,
-      fontSize: "1.105rem",
-      fontWeight: 400,
-      lineHeight: 1,
-      marginTop: 0,
-      marginBottom: 0,
-    },
-  },
-})
-
-export const printBreadcrumbFirst = style({
-  "@media": {
-    [mediaQueries.print]: {
-      fontFamily: printFonts.quattrocento,
-      fontSize: "1.105rem",
-      fontWeight: 400,
-      lineHeight: 1,
-    },
-  },
-})
-
-export const printHeader = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      textAlign: "center",
-      breakAfter: "avoid",
-    },
-  },
-})
-
-export const printDocument = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      paddingTop: vspace.one,
-      paddingBottom: 0,
-      fontSize: "10pt",
-      width: "100%",
-    },
-  },
-})
-
-export const printBlankPage = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "flex",
-      justifyContent: "center",
-      alignItems: "center",
-      breakBefore: "page",
-      height: "100vh",
-      fontStyle: "italic",
-      fontFamily: "serif",
-      fontSize: "1rem",
-    },
-  },
-})
-
-globalStyle(`${printDocument} button`, {
-  display: "none",
-})
-
-globalStyle(`${printDocument} h1`, {
-  "@media": {
-    [mediaQueries.print]: {
-      marginBottom: 0,
-    },
-  },
-})
-
-globalStyle(`${printDocument} h1 + ul`, {
-  "@media": {
-    [mediaQueries.print]: {
-      marginTop: 0,
-      marginBottom: 0,
-    },
-  },
-})
-
-globalStyle(`${printDocument} ul li:last-child::after`, {
-  "@media": {
-    [mediaQueries.print]: {
-      content: `""`,
-      padding: 0,
-    },
-  },
-})
